requirements.txt
fastapi==0.119.1
python-telegram-bot==22.5
google-generativeai==0.8.5
gspread==6.2.1
google-auth==2.41.1
google-auth-oauthlib==1.2.2
python-dotenv==1.0.0
uvicorn==0.38.0

